move useWaitForAsync to hooks.js
